+++ WinMSFactory/Admin/ProgramForm.cs
@@ -27,7 +27,7 @@ namespace WinMSFactory
 			{
 				dataGridViewControl1.IsAllCheckColumnHeader = true;
 				dataGridViewControl1.AddNewColumns("프로그램 코드", "PROG_ID", 100, false);
-				dataGridViewControl1.AddNewColumns("모듈", "MODULE_ID", 100);
+				dataGridViewControl1.AddNewColumns("모듈", "MODULE_NAME", 100);
 				dataGridViewControl1.AddNewColumns("프로그램 명칭", "PROG_NAME", 100);
 				dataGridViewControl1.AddNewColumns("사용여부", "PROG_USE", 100);
 				dataGridViewControl1.AddNewColumns("순번", "PROG_SEQ", 100);
@@ -48,16 +48,27 @@ namespace WinMSFactory
 		{
 			dt = programService.GetAllPrograms();
 			dataGridViewControl1.DataSource = dt;
+			ProgClear();
 		}
 
-		private void button1_Click(object sender, EventArgs e)
+		private void Search(object sender, EventArgs e)
 		{
+			if (((MainForm)this.MdiParent).ActiveMdiChild == this)
+			{
+				dt.CaseSensitive = false;
+				DataView dv = dt.DefaultView;
+				string search = txtSearch.Text.Trim();
 
+				if (search.Length > 0)
+					dv.RowFilter = $"PROG_NAME LIKE '%{search}%'";
+				else
+					dv.RowFilter = "";
+			}
 		}
 
-		private void button2_Click(object sender, EventArgs e)
+		private void Add(object sender, EventArgs e)
 		{
-			EmployeeVO employeeVO = new EmployeeVO { Employee_id = "A" };
+			EmployeeVO employeeVO = this.GetEmployee();
 			ProgramPopupForm frm = new ProgramPopupForm(employeeVO);
 
 			if (frm.ShowDialog() == DialogResult.OK)
@@ -66,25 +77,41 @@ namespace WinMSFactory
 			}
 		}
 
-		private void button3_Click(object sender, EventArgs e)
+		private void Delete(object sender, EventArgs e)
 		{
-			try
+			if (((MainForm)this.MdiParent).ActiveMdiChild == this)
 			{
-				string prog_id = dataGridViewControl1.GetCheckIDs("PROG_ID");
-
-				if (string.IsNullOrEmpty(prog_id))
-					return;
-
-				if (programService.DeleteProgram(prog_id))
+				try
 				{
-					MessageBox.Show("정상적으로 삭제되었습니다.");
-					LoadData();
+					string prog_id = dataGridViewControl1.GetCheckIDs("PROG_ID");
+
+					if (string.IsNullOrEmpty(prog_id))
+						return;
+
+					if (programService.DeleteProgram(prog_id))
+					{
+						MessageBox.Show("정상적으로 삭제되었습니다.");
+						LoadData();
+					}
+				}
+				catch (Exception err)
+				{
+					MessageBox.Show(err.Message);
 				}
 			}
-			catch (Exception err)
+		}
+
+		private void Clear(object sender, EventArgs e)
+		{
+			if (((MainForm)this.MdiParent).ActiveMdiChild == this)
 			{
-				MessageBox.Show(err.Message);
+				LoadData();
 			}
+		}
+
+		private void ProgClear()
+		{
+			this.Clear();
 		}
 
 		private void dataGridViewControl1_CellDoubleClick(object sender, DataGridViewCellEventArgs e)
@@ -92,7 +119,7 @@ namespace WinMSFactory
 			if (e.RowIndex < 0)
 				return;
 
-			EmployeeVO employeeVO = new EmployeeVO { Employee_id = "A" };
+			EmployeeVO employeeVO = this.GetEmployee();
 			int prog_id = dataGridViewControl1["PROG_ID", e.RowIndex].Value.ToInt();
 			ProgramPopupForm frm = new ProgramPopupForm(employeeVO, prog_id);
 
@@ -100,6 +127,12 @@ namespace WinMSFactory
 			{
 				LoadData();
 			}
+		}
+
+		private void txtProg_name_KeyPress(object sender, KeyPressEventArgs e)
+		{
+			if (e.KeyChar == 13)
+				Search(null, null);
 		}
 	}
 }
